// 分页

+++ sTraing/app/src/main/java/songming/straing/model/MissionInfo.java
@@ -8,41 +8,39 @@ import java.util.List;
  */
 public class MissionInfo implements Serializable {
 
+
     /**
      * trainingID : 9
      * userID : 10000000
      * type : 1
      * beginAt : 1461470371000
-     * finishAt : 0
+     * finishAt : 1461507554000
      * presetCount : 80
      * presetGroup : 8
      * perBreakTime : 30
      * location : 1
      * drinking : 1
      * gear : 1
-     * actualCount : 0
-     * actualGroup : 0
-     * actualBreakTime : 0
-     * actualConsumTime : 0
-     * status : 1
+     * actualCount : 80
+     * actualGroup : 10
+     * actualBreakTime : 300
+     * actualConsumTime : 500
+     * status : 3
      * createAt : 1461507067000
-     * content : 运动目标:
-     * 在 8 组内完成 80 个 伏地挺身;
-     * 由于你在 室外/操场 环境中，并平均休息 30 秒;
-     * 同时你选择使用 NIKE 护具和 肌肉科技 作为能量补充;
-     * 根据你实际的完成结果来最终评定你的得分
-     * title : 伏地挺身 #8G-80
-     * isSuccess : 0
-     * items : []
-     *
-     * training.status 状态：1.创建 2.开始 3.结束 training.isSuccess 0:未达成目标 1:达成
+     * content : 运动预设目标:
+     * 在 8 内完成 80 个 伏地挺身;
+     * 选择 室外/操场 环境，计划平均休息 30 秒;
+     * 选择使用 NIKE 护具和 肌肉科技 作为能量补充;
+     * <p/>
+     * title : 伏地挺身 #8G-640
+     * isSuccess : 1
      */
 
-    public long trainingID;
+    public int trainingID;
     public int userID;
     public int type;
     public long beginAt;
-    public int finishAt;
+    public long finishAt;
     public int presetCount;
     public int presetGroup;
     public int perBreakTime;
@@ -60,7 +58,7 @@ public class MissionInfo implements Serializable {
     public int isSuccess;
     public List<MissionProgress> items;
 
-    public static class MissionProgress{
+    public static class MissionProgress {
         public int count;
         public int breakTime;
     }
